backend/prompt_engine/requirements.txt
openai==0.28
python-dotenv
google-generativeai